Reduce # of snapshots captured on stream client
Stream clients only take 1 snapshot, but doesn't properly return
timespec start_time.
However, tries to take snapshots too many times, reducing throughput on
udp_stream significantly.
Reduce it so only 2 snapshots are needed.

author: lixiaoyan@
